Ignite Junction works
I need to add more tests and do some cleanup for this module.
